fixed user favorite

+++ app/src/main/java/com/aysesenses/n11_casestudy/ui/viewmodel/UserListViewModel.kt
@@ -1,21 +1,23 @@
 package com.aysesenses.n11_casestudy.ui.viewmodel
 
 import android.util.Log
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.*
 import com.aysesenses.data.firebase.FirestoreRepository
 import com.aysesenses.data.local.entitiy.UserEntity
 import com.aysesenses.data.local.repository.RoomRepository
 import com.aysesenses.data.network.api.GithubApiService
+import com.google.firebase.firestore.auth.User
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class UserListViewModel @Inject constructor(private val repository: RoomRepository, private val firestoreRepository: FirestoreRepository, private  val githubApiService: GithubApiService) :
+class UserListViewModel @Inject constructor(
+    private val repository: RoomRepository,
+    private val firestoreRepository: FirestoreRepository,
+    private  val githubApiService: GithubApiService,
+) :
     ViewModel() {
 
     // Internally, we use a MutableLiveData, because we will be updating the List of UserEntitiy
@@ -29,12 +31,12 @@ class UserListViewModel @Inject constructor(private val repository: RoomReposito
     private val userEntities: MutableList<UserEntity> = mutableListOf()
     private val mutableMap: MutableMap<String?, Any?> = mutableMapOf()
 
+
     fun userSearch(term: String) {
+        loadFromCache(term)
         viewModelScope.launch {
-            loadFromCache(term)
             val getPropertiesDeferred = githubApiService.searchUser(term)
             try {
-
                 //A successful response should include  returned all users with response.body()
                 val result = getPropertiesDeferred.body()
 
@@ -52,10 +54,6 @@ class UserListViewModel @Inject constructor(private val repository: RoomReposito
                 }
 
                 updateSearchResults(userEntities, term)
-
-                //_users.value = repository.getSearchResults(term)
-
-                //Save all users in Firebase
                 firestoreRepository.saveSearchResults(mutableMap,term)
             } catch (e: Exception) {
                 Log.e("userListErr", e.message.toString())
@@ -63,36 +61,30 @@ class UserListViewModel @Inject constructor(private val repository: RoomReposito
         }
     }
 
-    //This methot to favorite users
+    //This method to favorite users
     fun favorite(login: String){
         viewModelScope.launch {
             val list = repository.getUserFavoriteStatus(login)
-            Log.e("before",list[0].favorite.toString())
-
-            if (list[0].favorite == "no"){
-                repository.addFavorite(login)
-                Log.e("add","add favorite")
-            }else{
-                repository.removeFavorite(login)
-                Log.e("remove","remove favorite")
+            if (list.isNotEmpty()){
+                if (list[0].favorite == "no"){
+                    repository.addFavorite(login)
+                }else{
+                    repository.removeFavorite(login)
+                }
+                loadFromCache(list[0].term.toString())
             }
-            val list2 = repository.getUserFavoriteStatus(login)
-            Log.e("after",list2[0].favorite.toString())
-
-            //!!!
-            loadFromCache(list[0].term.toString())
         }
 
     }
 
-
     private fun updateSearchResults(userEntities: List<UserEntity>, term: String) {
         viewModelScope.launch(Dispatchers.IO) {
-            val favs = repository.getUserFavoriteStatus(term)
-            clearSearchResults(term)
+            val favs = repository.getFavorites(term)
             repository.insertSearchResults(userEntities)
-            favs.forEach {
-                favorite(it.login.toString())
+            if (favs.isNotEmpty()){
+                favs.forEach {
+                    favorite(it.login.toString())
+                }
             }
             loadFromCache(term)
         }
@@ -102,22 +94,11 @@ class UserListViewModel @Inject constructor(private val repository: RoomReposito
     private fun loadFromCache(term: String) {
         viewModelScope.launch() {
             val list = repository.getSearchResults(term)
-            _users.value = list
+            if (list.isNotEmpty()){
+                _users.value = list
+            }else{
+                Log.e("boş","boş dürüm")
+            }
         }
-    }
-
-    //Deletes all results in the database
-    private fun clearSearchResults(term: String) {
-        viewModelScope.launch(Dispatchers.IO) {
-            repository.deleteSearchResults(term)
-        }
-    }
-
-    private fun getFavorites(term: String): List<UserEntity>? {
-        var favs: List<UserEntity>? = null
-        viewModelScope.launch {
-            favs = repository.getFavorites(term)
-        }
-        return favs
     }
 }
